[android] #4365 - Moving example activities to more descriptive subpackage names for easier discovery
