+++ platform/android/MapboxGLAndroidSDKTestApp/src/androidTest/java/com/mapbox/mapboxsdk/testapp/espresso/DirectionsActivityTest.java
@@ -5,7 +5,7 @@ import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
 
-import com.mapbox.mapboxsdk.testapp.activity.api.DirectionsActivity;
+import com.mapbox.mapboxsdk.testapp.activity.directions.DirectionsActivity;
 import com.mapbox.mapboxsdk.testapp.R;
 
 import org.junit.Before;
