+++ platform/android/MapboxGLAndroidSDKTestApp/src/main/java/com/mapbox/mapboxsdk/testapp/activity/maplayout/MapPaddingActivity.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxsdk.testapp.activity.other;
+package com.mapbox.mapboxsdk.testapp.activity.maplayout;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
